src/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.0"
description = "Gota challenge"
authors = ["Martin Galdeano Cañizares<martingaldeca@gmail.com>"]

[tool.poetry.dependencies]
# Common
python = "^3.10"
setuptools = "61.1.1"
ipython = "*"

# Django
Django = "*"
Pillow = "*"

# Gunicorn
gunicorn = "*"
httptools = "*"

# Django jet
django-3-jet = "*"

# PostgreSQL
psycopg2 = "*"

# Utils
django-extensions = "*"
django-debug-toolbar = "*"
django-simple-history = "*"
django-model-utils = "*"
drf-api-logger = "*"

# Make sure static file is well configured
whitenoise = "*"

# Used to compress static / versioning
django_compressor = "*"
csscompressor = "*"

# Testing and QA
flake8 = "*"
pytest-django = "*"
factory_boy = "*"
ipdb = "*"
freezegun = "*"
pytest-xdist = "*"

# Django rest framework
djangorestframework = "*"
django-cors-headers = "*"
djangorestframework-simplejwt = "*"
drf-spectacular = "*"
user-agents = "^2.2.0"

# Celery
celery = "*"
requests = "*"
kombu = "*"
django-celery-beat = "*"
future = "*"

# Cache
django-redis = "*"

# Waffle
django-waffle = "*"
django-axes = "^5.23.0"
drf-extra-fields = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
